[change] Move AccountingView to freeradius endpoints #248
Closes #248
